requirements.txt
kubernetes
gin-config